// ts/react/common/types/bedev1Types.ts
export type TFriendVisits = {
	userId: number;
};

type TMediaAssetInfo = {
	wideImageAssetId?: string;
	wideImageListId?: string;
};

export type TMediaLayoutData = {
	primaryMediaAsset?: TMediaAssetInfo;
};

export enum TLayoutComponentType {
	TextLabel = "TextLabel",
}

export type TGameTileTextFooter = {
	type: TLayoutComponentType.TextLabel;
	text: {
		textLiteral: string;
	};
};

type TGameTileFooter = TGameTileTextFooter;

export type TGameTilePillComponent = {
	types: string[];
};

export type TGameTilePillData = {
	id: string;
	text?: string;
	icons?: string[];
	animationClass?: string | null;
};

export type TLayoutMetadata = {
	footer?: TGameTileFooter;
	tileBadgesByPosition?: TTileBadgesByPosition;
	pill?: TGameTilePillComponent;
	title?: string;
	primaryMediaAsset?: TMediaAssetInfo;
};

export type TTileBadgesByPosition = {
	ImageTopLeft?: TTileBadge[];
};

export enum TGameTileBadgeType {
	Text = "Text",
	Icon = "Icon",
}

export type TTileBadge = {
	analyticsId: string;
	tileBadgeType: TGameTileBadgeType;
	text?: string;
	icons?: string[];
	isShimmerEnabled?: boolean;
};

type TLayoutDataBySort = {
	[topicId: string]: TLayoutMetadata;
};

export type TGameData = {
	totalUpVotes: number | undefined;
	totalDownVotes: number | undefined;
	universeId: number;
	name: string;
	placeId: number;
	rootPlaceId?: number;
	playerCount: number;
	isSponsored?: boolean;
	nativeAdData?: string;
	isShowSponsoredLabel?: boolean;
	creatorName: string;
	creatorType: string;
	creatorId: number;
	creatorHasVerifiedBadge?: boolean;
	friendActivityTitle?: string;
	friendVisitedString?: string;
	minimumAge?: number;
	ageRecommendationDisplayName?: string;
	friendVisits?: TFriendVisits[];
	primaryMediaAsset?: TMediaAssetInfo;
	defaultLayoutData?: TLayoutMetadata;
	layoutDataBySort?: TLayoutDataBySort;
	navigationUid?: string;
};

export type TPresence = {
	gameId?: string;
	lastLocation: string;
	lastOnline: string;
	placeId?: number;
	placeUrl?: string;
	rootPlaceId?: number;
	universeId?: number;
	userId: number;
	userPresenceType: (typeof Presence.PresenceTypes)[keyof typeof Presence.PresenceTypes];
};
// ...
export type TGetFriendsResponse = {
	isOnline: boolean;
	isDeleted: boolean;
	description?: string;
	created: string;
	isBanned: boolean;
	id: number;
	name: string;
	displayName: string;
	presence?: TPresence;
	profileUrl?: string;
};