// ts/react/common/types/bedev2Types.ts
export enum TComponentType {
	AppGameTileNoMetadata = "AppGameTileNoMetadata",
	GridTile = "GridTile",
	EventTile = "EventTile",
	InterestTile = "InterestTile",
	ExperienceEventsTile = "ExperienceEventsTile",
}